refactor: separate normalised-task-status-chart
