+++ src/app/units/states/analytics/directives/charts/task-status-summary-charts/task-status-summary-charts.component.html
@@ -11,7 +11,7 @@
     </mat-card-header>
     <div class="flex items-center gap-2 p-4">
       <mat-form-field appearance="outline" subscriptSizing="dynamic">
-        <mat-select [(ngModel)]="campusFilter" (selectionChange)="onCampusFilterChange()">
+        <mat-select [(ngModel)]="campusFilter">
           <mat-option value="all">All</mat-option>
           @for (campus of campuses; track campus) {
             <mat-option [value]="campus">{{ campus }}</mat-option>
@@ -28,34 +28,12 @@
       </button>
     </div>
   </div>
-  @if (hasChartData) {
-    <div class="px-4 pt-8 text-slate-600">
-      <h3 class="mb-0">{{ selectedSnapshotDate }}</h3>
-      <p>Student count: {{ snapshotStudentCount }}</p>
-    </div>
-    <div class="h-125">
-      <ng-charts-bar-vertical-normalized
-        class="reverse-legend-order"
-        [animations]="true"
-        [gradient]="false"
-        [legend]="true"
-        [results]="data"
-        [scheme]="$any(colorScheme)"
-        [showXAxisLabel]="true"
-        [showYAxisLabel]="true"
-        [tooltipDisabled]="false"
-        [xAxis]="true"
-        [xAxisLabel]="normalisedCompletionSnapshotXLabel"
-        [yAxis]="true"
-        [yAxisLabel]="normalisedCompletionSnapshotYLabel"
-      >
-      </ng-charts-bar-vertical-normalized>
-    </div>
-  } @else {
-    <div class="flex h-full min-h-148 items-center justify-center">
-      <mat-progress-spinner mode="indeterminate" [diameter]="48"></mat-progress-spinner>
-    </div>
-  }
+  <f-normalised-task-status-chart
+    [campusFilter]="campusFilter"
+    [selectedSnapshotIndex]="sliderSelect"
+    [snapshots]="snapshots"
+    [unit]="unit"
+  />
   <div class="px-4 pt-2 pb-8">
     <div class="relative w-full">
       <mat-slider
